fix: `StripeController` test with too high test balance
Analysis shows that the `seedStripeDeposits` can/will create 3
valid Stripe deposits resulting in a balance of 117 euros. A top-up
of another 39 euros will thus fail as this exceeds the 150 euro
balance limit.

Instead of changing the seeding (which may inadvertantly affect
other tests) we simply change the amount of the top-up to €19.50.