feat(database): enforce SSL for database connection in production
Several ways devised to implement this easily. Most logical is to enforce this
feature only when actually running in production, but this does not work well
when using CI for testing. So a `TYPEORM_SSL_ENABLED` environment variable
provides more configuration flexibility (so you can also test this locally).

Originally, I wanted to do it this way:

```typescript
const sslConfig: any = process.env.TYPEORM_SSL === 'production' ? {
  ssl: {
    ca: fs.readFileSync('/etc/ssl/certs/ca-certificates.crt')
  }
} : {};
```

I thought about the potential complications of having `ssl: {},` but I do not
think it will cause any issues. I did try an alternative (smart) solution, but
the type checker does not agree :'( Even when casting
`as MysqlConnectionOptions | PostgresConnectionOptions` it does not work because
it starts complaining about the fact that the `ssl` property is `readonly`. So
it must be inlined.

Originally, the path to the CA certificates file was hardcoded. However, this
does not allow for flexibility on non-standard platforms that store it in a
different location or specific requirements from the environment. As such, the
`TYPEORM_SSL_CACERTS` has been introduced to handle that.