fix: balance cache invalidation on transaction creation
Currently, balance cache is cleared when a transaction is created/updated/
deleted for a user. However, balance cache should not be invalidated when
creating a transaction.

Creating a transaction is an action forward; new balance is just the sum of
balance cache + all new transactions for that user. Thus, no need to remove it.
Invalidating it is also bad for performance, as balance has to be calculated on
the fly until next balance cache calculation (cronjob).

This removes balance cache invalidation for transaction creations.